Update syntax
R is a bit tricky since there are technically no scalars. I am leaning
towards expressing generic lists as: `forall a . vector a`. In R,
heterogenous lists ("list") as well as homogenous lists are all vectors.
So an integer list would be `vector integer`. This is a little awkward,
since `integer` in R corresponds by itself to `[Int]` in Haskell. Or
rather, since missing data is allowed, it corresponds to `[Maybe Int]`.