Refactor as conventional R package
Eventually all morloc language modules should integrate with their
respective language communities. R modules should be valid R packages
and make use of all the ammenities available in the R ecosystem. They
should be pushed to cran or ropensci, same as normal R packages. The
`morloc` metadata should be passed neatly along in a way that is
accessible to the `morloc` compiler but not intrusive in the package.

I've also started a test suite. Morloc will have an testing as well that
may overlap considerably, but I think it is reasonable for tests to be
on both sides. Especially for serialization.