Make plot functions return NULL
I was encountering a type error in my old implementation of
mlc_plotVectorPDF. The function was suppose to return type '()', the
NULL type, from the IO operation. This will probably be a reoccuring
trend in R, that the R users will need to be more mindful of return
types than they might otherwise be.

I still need to hash out how IO is handled in morloc. Monads?